--- infra/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'groovy'
-    id 'project.laflex.library-mvc-conventions'
-}
-
-dependencies {
-    implementation project(':common')
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    runtimeOnly 'com.h2database:h2:2.2.224'
-}
-
-test {
-    useJUnitPlatform()
-}
-
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
