+++ infra/jpa/src/main/java/project/laflex/infra/jpa/config/JpaConfig.java
@@ -1,0 +1,15 @@
+package project.laflex.infra.jpa.config;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import project.laflex.infra.jpa.model.JpaModel;
+
+@EnableJpaAuditing(auditorAwareRef = "jpaEntityAuditAware")
+@EnableJpaRepositories(basePackageClasses = {JpaModel.class})
+@EntityScan(basePackageClasses = {JpaModel.class})
+@Configuration
+public class JpaConfig {
+
+}
