+++ infra/jpa/src/main/java/project/laflex/infra/jpa/model/JpaModel.java
@@ -1,0 +1,5 @@
+package project.laflex.infra.jpa.model;
+
+public interface JpaModel {
+
+}
