+++ infra/r2dbc/src/main/java/project/laflex/infra/r2dbc/InfraR2dbcApplication.java
@@ -1,0 +1,15 @@
+package project.laflex.infra.r2dbc;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.reactive.config.EnableWebFlux;
+
+@EnableWebFlux
+@SpringBootApplication(scanBasePackages = "project.laflex.infra.r2dbc.*")
+public class InfraR2dbcApplication {
+
+  public static void main(String[] args) {
+
+    SpringApplication.run(InfraR2dbcApplication.class, args);
+  }
+}
