+++ infra/r2dbc/src/main/java/project/laflex/infra/r2dbc/model/R2dbcModel.java
@@ -1,0 +1,5 @@
+package project.laflex.infra.r2dbc.model;
+
+public interface R2dbcModel {
+
+}
