+++ infra/r2dbc/src/main/java/project/laflex/infra/r2dbc/model/member/MemberRepository.java
@@ -1,0 +1,9 @@
+package project.laflex.infra.r2dbc.model.member;
+
+import org.springframework.data.repository.reactive.ReactiveCrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface MemberRepository extends ReactiveCrudRepository<MemberEntity, Long> {
+
+}
