+++ infra/r2dbc/src/main/resources/data.sql
@@ -1,0 +1,6 @@
+INSERT INTO member (id, created_at, created_by, last_modified_at, last_modified_by, username, phone)
+    VALUES (default, now(), 'System', now(), 'System', 'Mike', '010-1234-5678');
+INSERT INTO member (id, created_at, created_by, last_modified_at, last_modified_by, username, phone)
+VALUES (default, now(), 'System', now(), 'System', 'Jack', '010-1234-9876');
+INSERT INTO member (id, created_at, created_by, last_modified_at, last_modified_by, username, phone)
+VALUES (default, now(), 'System', now(), 'System', 'Kamus', '010-4321-5678');
