--- infra/src/main/java/project/laflex/infra/jpa/JpaModel.java
@@ -1,5 +1,0 @@
-package project.laflex.infra.jpa;
-
-public interface JpaModel {
-
-}
