--- infra/src/main/java/project/laflex/infra/jpa/member/MemberRepository.java
@@ -1,9 +1,0 @@
-package project.laflex.infra.jpa.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-
-}
