+++ settings.gradle
@@ -2,4 +2,11 @@ rootProject.name = 'laflex'
 include 'webflux'
 include 'common'
 include 'infra'
+include 'infra:jpa'
+findProject(':infra:jpa')?.name = 'jpa'
+include 'infra:r2dbc'
+findProject(':infra:r2dbc')?.name = 'r2dbc'
+include 'main'
+include 'infra:main'
+findProject(':infra:main')?.name = 'main'
